+++ src/jooom/chapter02/discount/condition/DiscountCondition.java
@@ -1,4 +1,6 @@
-package jooom.chapter02;
+package jooom.chapter02.discount.condition;
+
+import jooom.chapter02.Screening;
 
 public interface DiscountCondition {
     boolean isSatisfiedBy(Screening screening);
